fin_pjt/back/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
dj-rest-auth==7.0.1
Django==4.2.21
django-allauth==65.2.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
exceptiongroup==1.3.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
multidict==6.4.4
openai==0.28.1
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
python-decouple==3.8
python-dotenv==1.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
